+++ docs/releases.md
@@ -25,6 +25,7 @@ changes over time.
 * github-251: Support for log4j2 Configuration
 * github-252: Move sftp target into separate plugin
 * github-253: SQL Server relation should support explicit staging table
+* github-254: Use DATETIME2 for timestamps in MS SQL Server
 
 
 ### Version 0.27.0 - 2022-09-09